# docs/openapi-baseline.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Assistants Co Spine",
    "version": "0.1.0"
  },
  "paths": {
...
    "/hitl/queue": {
      "get": {
        "tags": [
          "hitl"
        ],
        "summary": "Get Queue",
        "operationId": "get_queue_hitl_queue_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HitlRequestRead"
                  },
                  "title": "Response Get Queue Hitl Queue Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/hitl/claim/{request_id}": {
      "post": {
        "tags": [
          "hitl"
        ],
        "summary": "Claim Request",
        "operationId": "claim_request_hitl_claim__request_id__post",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          },
          {
            "name": "agent_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/hitl/resolve": {
      "post": {
        "tags": [
          "hitl"
        ],
        "summary": "Resolve Request",
        "operationId": "resolve_request_hitl_resolve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HitlDecisionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/orgs": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Create Organization",
        "operationId": "create_organization_api_v1_admin_orgs_post",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/tenants": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Create Tenant",
        "operationId": "create_tenant_api_v1_admin_tenants_post",
        "parameters": [
          {
            "name": "org_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Org Id"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "tier",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/TenantTier",
              "default": "SOLO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tenant"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/tenants/{tenant_id}/suspend": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Suspend Tenant",
        "operationId": "suspend_tenant_api_v1_admin_tenants__tenant_id__suspend_post",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tenant"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/orgs/{org_id}/hierarchy": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Hierarchy",
        "operationId": "get_hierarchy_api_v1_admin_orgs__org_id__hierarchy_get",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Org Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/billing/{tenant_id}/ledger": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Tenant Ledger",
        "operationId": "get_tenant_ledger_api_v1_admin_billing__tenant_id__ledger_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HitlDecisionCreate": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "outcome": {
            "$ref": "#/components/schemas/HitlState"
          },
          "modified_draft": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modified Draft"
          },
          "feedback_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback Notes"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "agent_id",
          "outcome"
        ],
        "title": "HitlDecisionCreate"
      },
      "HitlRequestRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "work_item_id": {
            "type": "string",
            "title": "Work Item Id"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "state": {
            "$ref": "#/components/schemas/HitlState"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "claimed_by_agent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claimed By Agent Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "work_item_id",
          "reason",
          "state",
          "created_at",
          "claimed_by_agent_id"
        ],
        "title": "HitlRequestRead"
      },
      "HitlState": {
        "type": "string",
        "enum": [
          "PENDING",
          "CLAIMED",
          "RESOLVED",
          "REJECTED"
        ],
        "title": "HitlState"
      },
...
      "Organization": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "Organization"
      },
...
      "Tenant": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid",
            "title": "Organization Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "tier": {
            "$ref": "#/components/schemas/TenantTier",
            "default": "SOLO"
          },
          "status": {
            "$ref": "#/components/schemas/TenantStatus",
            "default": "ACTIVE"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "organization_id",
          "name"
        ],
        "title": "Tenant"
      },
      "TenantStatus": {
        "type": "string",
        "enum": [
          "ACTIVE",
          "SUSPENDED",
          "ARCHIVED"
        ],
        "title": "TenantStatus"
      },
      "TenantTier": {
        "type": "string",
        "enum": [
          "SOLO",
          "PRO",
          "BUSINESS",
          "ENTERPRISE"
        ],
        "title": "TenantTier"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}